+++ packages/component/src/Activity/CarouselLayout.js
@@ -33,7 +33,7 @@ const CarouselLayout = ({ activity, children, timestampClassName }) => {
               <React.Fragment>
                 <Flipper
                   aria-label={leftLabel}
-                  blurFocusOnClick
+                  blurFocusOnClick={true}
                   className={classNames(carouselFlipperStyleSet + '', filmStyleSet.leftFlipper + '')}
                   mode="left"
                 >
@@ -41,7 +41,7 @@ const CarouselLayout = ({ activity, children, timestampClassName }) => {
                 </Flipper>
                 <Flipper
                   aria-label={rightLabel}
-                  blurFocusOnClick
+                  blurFocusOnClick={true}
                   className={classNames(carouselFlipperStyleSet + '', filmStyleSet.rightFlipper + '')}
                   mode="right"
                 >